backend/requirements.txt
fastapi
uvicorn
openai
numpy
faiss-cpu
pydantic-settings
python-dotenv
pytest
pinecone-client
